refactor: phase 1 service split — portfolio + strategy shim fixes
## portfolio_service.py (2165L) → 7 files + shims

Inheritance chain:
PortfolioBaseService (init, caches, thread pool, static helpers)
  -> PortfolioPriceService (price fetch, daily change, sector)
    -> PortfolioCoreService (holdings CRUD, P&L, summary, CSV)
      -> PortfolioArchiveService (portfolio snapshot archive)
        -> PortfolioExecutionService (record_sell, trade history)
          -> PortfolioRiskService (sell rules, presets, evaluation)
            -> PortfolioService (assembled in __init__.py)

New files (all under api/services/portfolio/):
- portfolio_base_service.py (185L)
- portfolio_price_service.py (319L)
- portfolio_core_service.py (699L)
- portfolio_archive_service.py (229L)
- portfolio_execution_service.py (144L)
- portfolio_risk_service.py (703L)
- portfolio_alert_service.py (474L) — absorbs portfolio_alert_scanner.py

Shims (backward compat):
- portfolio_service.py → 22-line re-export
- portfolio_alert_scanner.py → 24-line re-export
- portfolio_alert_service.py → 16-line re-export

## Strategy shim fixes
- Add _compute_inverse_vol_weights, _compute_risk_parity_weights
  to strategy/__init__.py and strategy_service.py shim
- Add ScreeningService to strategy_service.py namespace
  (required by monkeypatch in test_strategy_service.py)
